Set fuzzy after setting content
Otherwise the Po backend is too smart and drops fuzzy flag when the
string is not translated. So creating new fuzzy translation leads to
situation when in database we have the fuzzy flag but not in the
backend.

This inconsistency in database is causing wrong behaviour when togging
fuzzy flag as seen in issue #189.