Simplify handling of fulltext index
Use async writer rahter than buffered one - this way we loose some
performance in case index offloading is not enabled, but we are able to
work with index even in multi threaded environment out of the box.

If user wants performance, indexing offloading is anyway the only way to
go.

Fixes #289