Refuse to load .pot as base file
Previous using of template in admin interface lead many users to
specifying pot (PO template) as it, what is wrong as this (now called
base file) is meant only for monolingual formats.

There don't seems to be any legitimate use of pot file as base file, so
let's completely forbid this option.