Document new caching of Avatars
Avatars are now downloaded server side and cached for all users. This
improves performace of the site (no need to request to third party
servers) and also improves users privacy for the same reason.

Caching can be configured separately from rest of the Django caching and
it is recommended to use filesystem backed cache for this purpose.

Fixes #412

Signed-off-by: Michal Čihař <michal@cihar.com>