Use new language codes for Traditional and Simplified Chinese
Let's prefer new locale codes for these:

https://www.iana.org/assignments/lang-tags/zh-Hant

https://www.iana.org/assignments/lang-tags/zh-Hans

This also removes previous renaming of zh_TW and zh_CN as these are no
longer the primary locales for either of these.

We though don't rename the locale files as Django 1.6 does not yet
support this and we still support it (see issue #630).

Signed-off-by: Michal Čihař <michal@cihar.com>