Better handle Qt TS differences
Move the code to sepatate class, so that we do not have to check class
of unit.

Differentiate plurals, thouhg this is still far from perfect as singular
keeps (s). This is needed for message lookup to work as we would have to
store the original string somehow as well for the lookup and editing (or
make editing slower by doing the transforms while searching for unit to
update).

Issue #1271

Signed-off-by: Michal Čihař <michal@cihar.com>