Better detection of unmerged changes.
'heads(branch(.)) - .' revset finds heads of the current branch,
excluding the current changeset if it is a branch head.

Hence, it returns a non-empty set when:

 - there's a single branch head, we're not at it
   (and we can ff to it instead of merging or rebasing)
 - there are multiple branch heads

An empty set is returned iff there's only one branch head and
we're at it.

The original revset didn't ignore other branches as it supposed
the upstream is always at tip, which it isnt'.

Signed-off-by: Andrew Shadura <andrew@shadura.me>