Simplify file save code
We really don't need magic the translate-toolit save and safefile
methods provide - our filenames do not change and we just want to
setialize the file into current filename.

This also potentionally avoids problem in translate-toolkit which is not
using contexts for manipulating with files, leading to possibility of
corrupting output files in case of error.

Issue #1426

Signed-off-by: Michal Čihař <michal@cihar.com>