Share code for adding new translations
In the end all file formats behave quite similarly, so it makes little
sense to duplicate the same logic for all formats.

- if base file is specified, remove translations, set target language
- if no base is specified use file format specific template if available

This way we can better deal with file formats which generally start with
empty file, but in some cases users want to have all keys - now it
depends only if the base file is set or not.

See also #1672

Signed-off-by: Michal Čihař <michal@cihar.com>