Use temporary dirs instead of named temp files in tests
The current code was written based on assumption that files can be
opened multiple times, what is not true on Windows. Using
NamedTemporaryFile with delete=False is option as well, but that just
makes the code too complex, having separate dir where the tests happens
make it simple and easy to reuse.

Signed-off-by: Michal Čihař <michal@cihar.com>