Avoid looking up string combinations in glossary
This really can not scale for large strings. Generating 30k of
match strings for 1kb text is simply too much. This consumes too much
computing power on the Weblate side and leads to problems on the
database side as it can not evaluate such regexps. Doing the exact match
for all the phrases didn't work as well see
555b1483ea4d18cd451ded7a17cd98cf0ca5f5d9.

Instead we now do simple substring lookup for all stemmed words. This
can produce slightly more matches, but is way faster.

The proper fix would be to have stems in the database and operate on
that. The problem here is that it would slow down other parts as well.

Fixes #1773
Fixes rb#1368

Signed-off-by: Michal Čihař <michal@cihar.com>