Move gitexport initial migration to later phase
We need to allow later installation of the app, but it has become
problematic after SubProject -> Component rename as the migration had to
be executed before rename, what is no longer achievable on new
installations.

Instead the initial migration is now empty and there is new migration
which does it's job and is executed after renaming.

On the other side changing depenendecies of initial migration and moving
it to Component would not work for existing installs - they would have
applied migration without its dependencies.

Fixes #2065

Signed-off-by: Michal Čihař <michal@cihar.com>