Add batch target check operation
For some checks it is better to perform them once after translations are
imported as they do heavy database queries which can be done more
effectively globally on the database side. The typical use case is
consistency check which triggers way too many queries when executed for
each unit, but it's relatively cheap when done globally.

Fixes #2318

Signed-off-by: Michal Čihař <michal@cihar.com>