Clarify supported states for XLIFF format
Previous documentation incorrectly mentioned "fuzzy" which is a concept that
does not exists in XLIFF and was also replaced in Weblate by "Needs editing".

More importantly it was very unclear that the `state` attribute is never used
at all. But instead only `approved` attribute is used. That lead to even more
confusion for the reader.

The following tables show all possibles cases in both directions:

## From XLIFF file to Weblate

| XLIFF file            | Weblate            |
|-----------------------|--------------------|
| absence of `<target>` | Empty              |
| absence of `approved` | Waiting for review |
| `approved="no"`       | Waiting for review |
| `approved="yes"`      | Approved           |

## From Weblate to XLIFF file

| Weblate            | XLIFF file       |
|--------------------|------------------|
| Needs editing      | `approved="no"`  |
| Waiting for review | `approved="no"`  |
| Approved           | `approved="yes"` |

This should address the questions raised in
https://github.com/WeblateOrg/weblate/issues/1040#issuecomment-433366326

Signed-off-by: Adrien Crivelli <adrien.crivelli@gmail.com>