Use own index for finding units
This way we do not have to check translate-toolkit results (which
sometimes do not match our expectations) and also we can have index on
both context and source making the lookups way faster.

This change leads to stricter API which needs adjustments in the tests
for some monolingual files used as bilingual.

Fixes #2249

Signed-off-by: Michal Čihař <michal@cihar.com>