Change anonoymous session expiry time
The original value of 30 minutes was not really a good choice as it
could clash with cron jobs scheduled to run every 30 minutes. The
problem is that if an anonymous client connects in exactly 30 minutes,
the request starts with still active session, but it expires until it is
completed and saving it fails. This leads to SuspiciousOperation "The
request's session was deleted before the request completed. The user may
have logged out in a concurrent request, for example." being raised by
Django.

Signed-off-by: Michal Čihař <michal@cihar.com>