Add a configuration option to make the translation length limit based on the source string optional.
Use case:

When translating language files which are key value based with short keys like "home.description" and a long translation Weblate currently limits the length of the translation to max(100, source.length * 10), in this example 160 chars. To avoid adding a manual max-length constraint to all keys tapping into this case the option LIMIT_TRANSLATION_LENGTH_BY_SOURCE_LENGTH can be set to false. If it's set to false the calculation will be skipped and instead translations up to 10.000 characters are allowed.

Signed-off-by: Enno Woortmann <enno.woortmann@web.de>