Add workaround for tesseract 4.0
It needs to have C locale for it's execution due to code relying on
locale dependant functions for parsing trained data (eg. sscanf).

This is really a workaround, but given that tesseract 4.0 is shipped
with upcoming Debian stable, we will have to live with this for quite
some time.

Fixes #2581

Signed-off-by: Michal Čihař <michal@cihar.com>