Better handle inconsistent changes with pending state
In case there is no recent change for pending unit, use anonymous author
and current timestamp. This way we can flush the pending units even when
the database seems to be wrong.

This can happen in case somebody manually sets pending flag or in case
of bug in our code.

Fixes #2590

Signed-off-by: Michal Čihař <michal@cihar.com>