Avoid using interpshinx to link to Weblate 3.0 docs
The intersphinx also resolve links without prefix and that makes
removed targets link to old docs. That way it is easy to miss that
something is still poiting to the old docs while it should not. Without
interpshinx for Weblate 3.0 we get build error in case of non existing
target.

Fixes #2630

Signed-off-by: Michal Čihař <michal@cihar.com>