Removed file format autodetection from component setup
This was not really a good design choice in the beginning - it is not
reliable and uses lot of guessing instead of asking user for thing which
should be easy to figure out. With translation-finder, Weblate discovers
the correct setup automatically during adding component and this makes
even less sense.

On the other side this slows down several operations as we need to parse
the file to figure out file format - now the authoritative information
on this will be already in the database.

Fixes #2660

Signed-off-by: Michal Čihař <michal@cihar.com>