Addons: Trigger language consistency daily
Running it post update is wrong because the translation files are not
parsed at that point. Also running it can be expensive and we do not
want to slow down the update process. Running it daily in the background
seems better choice as it will not slow down anything and will pickup
any possibly manually created translations once they are present in
Weblate.

Fixes #2741

Signed-off-by: Michal Čihař <michal@cihar.com>