Checks: Add Qt formats checks
The Qt library support formatting with %1, %2, etc ... placeholders and
the QString::arg() function.

Also plurals are supported in Qt with a special "%n" format. Weblate
already supports this in Qt TS files, but there was no support for
highlighting and checking these placeholders.

Like other format, the checks needs to be enabled with the qt-format and
qt-plural-format flags. GNU xgettext supports them with the special --qt
argument, but PO is not the preferred format for Qt.  Unfortunately, Qt
own tools do not add these flags, so they will need to be enabled
globally for the entire file.

Add checks, doc and tests for the 2 flags.

Signed-off-by: Vincent Tondellier <vincent+github@tondellier.com>