Checks: Add Ruby format checks
Ruby supports a dialect of the standard printf format, with some
extensions for named variables.

At the time of this commit, GNU gettext does not support Ruby as a
language, and ruby-gettext's xgettext has no support for flags. So
invent a new ruby-format flag that will need to be enabled for the
entire file.

Add checks, doc and tests.

Signed-off-by: Vincent Tondellier <vincent+github@tondellier.com>