Allow extending instead of replacing LOGIN_REQUIRED_URLS_EXCEPTIONS
Quite a few of the URLs included in `LOGIN_REQUIRED_URLS_EXCEPTIONS` by default are needed for Weblate to function properly. When using the Weblate docker container, if you need to add a single path as exception then you suddenly need to replace them all, which may cause issues if they change during an update.

By allowing the extension next to the replacement it's trivial to add a single URL while still allowing Weblate to work as normal.