Memory: Remove wiping support
It is no longer needed with in database storage. In case replacement is
needed, better to do this with proper UI or API (selection of what to
wipe). In the end this can now be done in the admin interface if really
needed.

The user or project entries will be removed on removal.