Auth: Improve permissions caching
Fetch all user permissiosn in single query. We typically check for bunch
of permissions in every request, so it's better to fetch them all at
once and then do cheap lookups in Python.

This replaces per function cache layer on individual permissions as it
is no longer needed as individual lookups are now cheap.

This is also prepartion for extending permission system in future.
Restrictions or component level permissions are now easier to implement.

Issue #3646
Issue #3267