Accounts: Use own middleware for activating locales
We activate the locale based on user profile and use cookie only as
auxiliary source (for example when user logged out). This avoids need to
hook the locale change in other places as the middleware now deals with
both setting the correct language on request start and when the settings
have changed or cookie is missing.

Fixes #3833