Unit: Remove state caching flags
Directly reference checks, comments or suggestions instead of relying on
cached flags. This was meant to improve the performance, but it doesn't
bring reasonable benefits once we have switched to using foreign keys
for linking these to units. On the other side the cost of updating this
is quite big, especially on components with lot of strings and failed
checks.

Fixes #3843