Translation: Bring detailed listing color indications in sync with progress
Try to be as close as possible, but there will never be a 1:1 mapping as
some of the items are not indicated in the progress (for example
comments or suggestions) and some might appear in more categories
(approved string with failing check).

Also use label colors to better differentiate the labels visually.

Fixes #4085