Middleware: Redirect to renamed projects/components
This makes it easier to change the URL slug, what seems to be popular
these days as some projects used "master" in URL slug to match their
main branch and they are renaming it now.

The case insensitive and language alias lookups are now handled by
redirects as well. This makes it clear what is the canonical URL and
makes the canonical URLs load faster as it no longer has to do case
insensitive lookups in the database.