API: Consistent error handling
Remove result from failed operations. That is error prone as in some
cases result is boolean value on success and making condition on result
attribute which is a string in case of failure results in false result
evaluation. The proper way to check status of operation is HTTP status
code.