Unit: Change referencing to source units
It is now explicit relation instead of relying on id_hash. This is more
robust and makes it easier to query on source string properties. The
Weblate metadata is now stored only on source unit avoiding need of
propagating them on every change making import, labeling of flags
updates faster.

Fixes #4190
Fixes #4020