Auth: Change way in which session expiry is set
Setting the session expiry in middleware has side effects of doing a
session change in every request, what is not desired. It also forces
session creation when there is no session needed.

The logic is now reverse - the sessions defaul to shorter livespan,
which is used for anonymous sessions and authenticated ones have
separate setting to set their validity which is set on login. The
session has to exist at this point and we are changing it just once per
session lifetime.

See https://hackerone.com/reports/1003468