Unit: Improve checks propagation, take two
The original approach was wrong as it did cut the recursion only in one
specific case, while it is problematic in others as well. Whenever
checks are executed, there was possibility that this would recurse for
every unit leading up to N*N invocations instead of just N. The impact
was depending on (random) database ordering, so it was not always
visible.

See #4661
Fixes WEBLATE-56E