UI: Simplify indicating incomplete languages
There is now ghost icon with tooltip indicating amount of not translated
strings. The confusing "of total" columns is now gone.

This way the most imporant information (there are some strings not
translated into this language) is visible right now and more detailed
info (how much strings) is in the tooltip.

Fixes #2866