UI: Make language in the translator clickable
This allows to drop the pencil icon and make the screen less cluttered.
The link is there now show in all cases as user can always access the
other translations and there is no reason to block this based on the
edit permission (he might be able to suggest there).

Issue #4697