ci: Switch back to using pre-commit integration over local
This allows to use pre-commit.ci service which can provide fixes to pull
requests, what is really nice to new contributors.

We loose tight control over versions of flake8 plugins. In case this
will cause troubles, we might introduce automation for that, but still
that would be less work than implementing the auto fixer.