translation: Remove parsing logic on save
There should be no changes upon saving. When they are this
most likely indicates a bug and parsing only some attributes
will just make it worse.

For example when saving does not untranslated previously translated
string, the state would be revereted, but not the content, producing
"translated" strings with empty translation.