alerts: Add alert to unused components
These typically can be removed if there is no activity going on.

For now it ignores glossary components (all projects get them, but some
might not use them) and fully translated components (it is impossible to
guess whether component is unused because it is dead upstream or because
there is nothing to translate).

See #9086