ui: Add approved percentage to the object listing
The new column is displayed in front of the translation percentage because
when approvals are active they probably become the main progress criteria.

On a smaller display resolution it will be shown in third priority after
the name and 'Translated' column.
The other columns appear 100px later than before, except
'Unfinished Characters' which appears after 'Comments' at an additional
100px instead of 200px. This allows 'Untranslated' to show at the usual
1400px, so on common 1440px wide screens no information is lost.

Fixes #9855