translation: skip strings which can not be serialized
When string serialization fails, skip the string, unmark it as pending
and continue in other pending strings. The error is stored as a Change
object and the string is marked as needing editing.

This makes it possible to recover from the situation, but the string can
be lost as it is not serialized, the change will disappear on next upstream
update.

Right now this happens frequently with Fluent, but can happen with other
formats as well.

Fixes #9775