stats: avoid fetching last change from database when doing it
Improve logic to store the last change object locally as well so that
update_stats() does not have to fetch it again. This needs to be done
immediatelly to be available right after creating the change (contrary
to what update_cache_last_change() does - it updates the cache for other
processes).