backups: use static files to download project backup
These can be huge and would block wsgi for long. Instead, we copy the
file temporarily to the static storage and redirect to it. This way the
download is handled directly by the web server, which can handle it more
effectively than wsgi and can also support resuming of interrupted
download. As a side effect, it can be donwloaded via command line tools
as the randomly generated temporary URL doesn't require authentication.