create_translations: Enable background processing by default.
Adds a new `run_async` parameter to `create_translations` function,
`False` by default.

When `False`, the update will processed immediately, and any failure
will be reported.

When `True` (and the debug `CELERY_TASK_ALWAYS_EAGER` setting is not
`True`), the update is send to a Celery task and will happen in the
background.

The `create_translations` calls from `add_new_language`, `do_update`,
`do_reset`, and `do_file_scan` are made asynchronous.

From a user PoV, it means that the execution of these tasks now returns
(very) fast, and the actual processing happens in the background, which
is trackable from the UI itself.

NOTE: This commit _may_ also fix a theoretical re-entrant loop? If the
call to `create_translations` from the Celery task failed again to
acquire the component lock, existing code would add a new task?