test(format): more realistic PHP keys for test_new_unit
Extracted from #11937 since that one is likely going to take a while
and IMO this is worth merging on its own.

The test data wasn't realistic for PHP files: It happened to work, but didn't create a correct PHP file. We fix the test fixture to have more realistic data.