fix(formats): do not copy template notes for new translations
Do not copy the notes (comments) from the template/source when adding
a translation for a key (in monolingual formats).

This fixes #11133. Copying the notes is problematic because the are
not synced when they are updated. As a result, they can get out of
date.

The implicit assumption behind this change is that most users only
care about comments in source files (but not in target files).

In addition to the unit tests added in this commit, I have tested
manually (with the dev weblate), that with these changes, new
translations in Java property files do not have comments anymore.