Feat: Add clipboard paste functionality for screenshots
The user clicks paste. Then the clipboard is read if the permesion is given and the context is secure (as required by the clipboard API). Then the image is pasted into the select file input. Otherwise if no image is available we only inform the  user about that.

NOTE: Implemented only in the add screenshot form in /screenshots page and not in the modal.

Related: #6710