fix(ui): Update suggestions box (#12129)
* Update suggestions box

The actions box no longer covers the text field under it. Also fixed spacing between the buttons inside the box.

Closes: #8356

* Update button styles in style-bootstrap.css

Bring back removed styles. They are used by several elements. If removed they break some of the UI.

Related: #8356

* Fix comments and suggestions tooltip

Add gap between suggestion and comment elements. To prevent the tooltip from covering the content of the item before it

* Fix RTL text alignment

Fixed the text to show completely on the right side of the suggestion view box. Instead of showing in the middle.

* chore: update gap in container-gapped to 40px