feat(screenshots): Add clipboard paste functionality for screenshots (#12181)
* Feat: Add clipboard paste functionality for screenshots

The user clicks paste. Then the clipboard is read if the permesion is given and the context is secure (as required by the clipboard API). Then the image is pasted into the select file input. Otherwise if no image is available we only inform the  user about that.

NOTE: Implemented only in the add screenshot form in /screenshots page and not in the modal.

Related: #6710

* Add screenshot clipboard paste in the modal box.

The previous commit added pase screenshot in screenshot/ pages. This adds it in the modal box that opens when Add Screenshot button clicked while checking individual strings.

Related: #6710

* Move the paste button closer to image input

The paste button now is rendered  by django screenshot form isntead of being in the template.

* Remove the paste button if clipboard API  isn't supported by the browser

* Remove part of the forgotten paste button

* Use custom input instead of using dummy one

* Add paste button to screenshot editing

* Add changelog entry

Co-authored-by: Michal Čihař <michal@cihar.com>