feat(ci): separate linkcheck to standalone workflow
The linkcheck is less and less reliable in past weeks. It seems that
some of the sites are broken often, it might be also caused by them
blocking GitHub ranges as such probing might be visible on some of them.

Separating the workflow allows us to execute it only on documentation
changes (we still want to run regular documentation build on all pull
requests to make it required).