feat: add keyboard shortcuts modal and functionality (#12430)
This commit adds a new keyboard shortcuts modal and binds it to "?" shortcut. The modal displays a table of available keyboard shortcuts and their corresponding actions. The keyboard shortcuts are implemented using the Mousetrap library. The modal is triggered by pressing the "?" key.

* updated the ID of the keyboard shortcuts modal to "shortcutsModal" for better accessibility


* Refactor keyboard shortcuts modal for better translation support
* Add translation context to shot phrases in keyboard-shortcuts.html
* Update keyboard-shortcuts.html
* small sentence corrections
* docs: document keyboard shortcut changes
* feat(keyboard-shortcuts): Add conditional check for input fields

This commit adds a conditional check in the keyboard-shortcuts.js file to prevent the shortcuts modal from opening when an input field or textarea is focused. The check ensures that the modal is only shown when the user presses the "?" key outside of input fields or textareas.

Fixes #5064

Co-authored-by: Benjamin Alan Jamie <benjamin@weblate.org>
Co-authored-by: Michal Čihař <michal@cihar.com>