feat: quote audit log values to avoid auto linking
Some e-mail services such as Gmail automatically convert things that
look like a link into links even in HTML mails. This could be
problematic for audit log entries as want to display the information as
is without turning that into links.

Adding <span> seems to help in this according to
https://stackoverflow.com/a/23404042/225718