fix(editor): correct row offset calculation in translation editor
The logic used to the calculate scroll offset based on the closest tbody tag which is fine unless the space between them is very large (like in issue #3507) which hides the translation input. As a fix we calculate based on a closer element that can't grow in height very much to hide the translation input.