chore(widgets): revamp widgets page
Revamp the widgets page. Use a card based layout for settings, preview and showing
the embed code. The current refactoring unifies the settings UI on the frontend
and generates the urls for various widgets on the frontend as well. Currently, this
serves no extra benefit but will prove to be helpful in future when we add additional
settings which are only applicable to specific widgets.