feat: Allow inspect followed project (#14452)
* Allow right click to inspect project

This allows right click in /accounts/profile/#notifications in the watched project list. So users can open a project and inspect it before follwing it or not, as was suggested in issue #12434

* Open project in new tab and add info label

* Add missing migeration

* Make languages elements link link

They behave exactly like a normal link (right/mid-clickable), but noraml click is used to select and unselect.

* Fix potential XSS vulnerability by encoding user input in href attribute

- Added  to sanitize the  before inserting it into the  attribute.
- Prevents unsafe HTML/JavaScript injection in URLs and mitigates potential DOM-based XSS attacks.

* Fix hidden watched project

As was noted in #13889

Close: #12434