fix(notifications): truncate large digest emails (#14513)
If there are a large number a notification in a digest, the digest is chunked into
multiple emails. This can result in a large number of emails for the user which is
annoying. Hence, only send the first 100 (current digest chunk size) notifications
to the user in the digest. If the number of notifications exceeds this, a note that
the digest was truncated is displayed to the user in the digest email.

Fixes #14130.

Co-authored-by: Michal Čihař <michal@cihar.com>