feat: import comment api
Update the /api/units/{unit_id}/comments/ endpoint to accept user and timestamp for a comment
if the user making the request is a project admin. Return the comment id in the response as
it is useful in testing and can also be useful in future resolve/delete comment APIs.