feat: sort units by component when translating multiple components
Add sort choice to sort by component and unit priority to avoid unnecessary jumping
between contexts. The units are sorted by the component in the same order as components
are sorted in a project/category.

Further, move the sort name and sort query to the SearchForm from the views to ensure
consistent sort defaults on all pages.