bugfix: set filter_pos to offset in zen mode
filter_pos sets the default offset for PositionSearchForm's offset field. The min value
of the offset should be 1 as get_zen_unitdata subtracts one from it. Passing it as 0
makes it -1 and that filters out all the results in it. This happens when the sort order
is changed in zen mode.

Passing in the current offset to filter_pos fixes the issue.