feat: sort units by component when translating multiple components (#14781)
Add sort choice to sort by component and unit priority to avoid unnecessary jumping
between contexts. The units are sorted by the component in the same order as components
are sorted in a project/category.

Further, move the sort name and sort query to the SearchForm from the views to ensure
consistent sort defaults on all pages.

filter_pos sets the default offset for PositionSearchForm's offset field. The min value
of the offset should be 1 as get_zen_unitdata subtracts one from it. Passing it as 0
makes it -1 and that filters out all the results in it. This happens when the sort order
is changed in zen mode.

Passing in the current offset to filter_pos fixes the issue.

Co-authored-by: Michal Čihař <michal@cihar.com>