feat(api): import comment api (#14734)
* feat: import comment api

Update the /api/units/{unit_id}/comments/ endpoint to accept user and timestamp for a comment
if the user making the request is a project admin. Return the comment id in the response as
it is useful in testing and can also be useful in future resolve/delete comment APIs.

* Refactor comment creation logic and API response

Enhanced API response to include serialized comment details.

* fix user = None in comment creation

* add missing migration