feat(api): enable project creation within billing limits
Added support for non-superusers to create projects if they have billing enabled and
the project limit is not exceeded. Currently, a within limit billing is automatically
used for the project. Since most users have only one billing, it should work fine for
most cases.

To let users with multiple billings choose the desired billing, the billing id or
a similar identifying field needs to be exposed to the users. Currently, the user
can only view the billing id from the url of the billing details page, which is not
exactly ideal.