chore: cleanup translation create interface
Historically there was `create_translations` which decided whether to
make the processing immediate or in a task based on the configuration
and parameter. Remove that parameter and make it always use background
processing if available. Also renames the immediate processing method to
make it clear what it does and use it where sync behavior is needed.