feat(memory): add option to disable personal and project TM contribution
Introduce `contribute_personal_tm` for user profiles and `contribute_project_tm` for components,
to enable selective contribution to translation memory. If `contribute_personal_tm` is disabled,
the user's translation memory is not updated. If `contribute_project_tm` is disabled for a
component, the associated project's translation memory is not updated.