fix(changes): Fix faulty test_user test in test_changes.
The test was testing the existence of a `Translation added` string
which is not unique to the user making the change. So we test for
`Hello, World!` instead. The existing test case would pass without
passing the filter to the request since the change made by the
current user would be part of the latest changes. To check whether
the filter is working as intended, we would need to check whether
filtering by another user will exclude that change. We add relevant
test code for that.