chore(pending changes): deepen the stack of changes
Currently, there are multiple situations under which translating a unit requires that
the previously pending changes to the unit be flushed to the repository. One such example
is two different users editing the same unit, and the second user's changes cannot be saved
till the first user's changes to the unit have been committed. This is a performance issue.

To fix this issue, introduce a new PendingUnitChange model that keeps track of the pending
changes to be flushed to the repository. When a unit is translated or saved otherwise, a
PendingUnitChange record is created for the change instead of flushing the changes to the
repository.

The logic to group changes by author has changed as the previous code depended on the changes
to a unit by a different author to have been already committed. The new logic handles these
cases of changes to the same unit by multiple authors correctly.